+++ test/test_helper.rb
@@ -1,5 +1,9 @@
 # frozen_string_literal: true
 
-require "minitest/autorun"
-require "yard/relative_markdown_links"
+require "fileutils"
+require "open3"
 require "ostruct"
+require "yard"
+require "yard/relative_markdown_links"
+
+require "minitest/autorun"